Redesigned backend to use a database, and moved flask endpoints to separate file. Updated facial recognition, and added object detection system.
All data is now stored in an SQLight database, instead of individual .fetula files. This will allow for faster data reading, easier upgrading and safer writing. Fixed the facial recognition backend, and added many features on the frontend, including a page showing all the photos a person is in. Options when viewing a photo include adding a person to an image, changing a detected person and deleting a detected person. Albums can be renamed, as can people.

Added object recognition to the backend, using Google's Tensorflow object detection api and the Imagenet model. Over 500 different objects can be detected in an image with reasonable accuracy. At the moment, the process is incredibly slow, much over a minute per photo. Importing the api also takes over two minutes, so the feature is completely removed when working on other features. The objects detected in the image are stored as 'tags' and are displayed on the view image page.

The endpoints for the webserver have been moved from server.py to api.py, as I am now using the application factory system for Flask. Created a lazy image loading component from scratch, but disabled most of its features. The main delay when it comes to image loading was the .fetula files, and that delay has now been eliminated, making image loading much faster.