Refactored scanning and processing. Added an 'AI Assistant', 'Athena'.
The user can now process images from the frontend, and decide which things to scan for. Processing now uses multi-processing, which for small operations makes no difference speed wise but is significantly faster for longer running scripts. Currently only the progress of the scan is shown on the frontend.

Perception hashes can be generated for images, and groups of similar images can be scanned for. Also added a library that can detect the rotation of an image, to see if its orientation is correct. The results of these scans are placed on the new 'Assistant' ('Athena') page. Animations can also be generated from burst shots, although this hasn't been perfected yet. A script to detect if an image is blurred has been added, but in initial tests it did not perform well (too many false positives).

Exif orientation is now applied to an image, and the exif orientation is reset. This makes serving images faster, and removes any problems that may arise with writing exif data, or opening the images in other apps. Added a file lock system to prevent multiple threads opening the same image during this process. When an image is rotated in the app, that rotation is actually carried out, not stored as an exif value.

Photos can now be deleted (moved to a bin) tagged or rotated from the frontend.

The lazy image loading system has been reworked and fixed, and now performs as intended. The aspect ratio of the image is sent with the metadata, and the placeholder/main version will only start loading when the page has been scrolled to a certain point.

Album thumbnails now show the size and date range of the album. The face highlighting component has been renamed to the 'detected' component, as it can now highlight objects as well as faces. Added a new sidebar to the view page with edit tools, at the moment just rotation. Images can also be rotated on an album page using controls in their thumbnail. Moved the tags editor into its own component. They can be removed and added. They can also be generated by pressing a button, which will spin while awaiting a response from the backend.

Added a scroll to top button. Also made the scrollbar visible, and made it look modern. Changed the style of the home page. The view can now be dynamically changed to replace the old albums view with album thumbnails. The assistant and process buttons are alongside the search bar, and the assistant icon has notifications next to it. Also added the logo to the toolbar.

Added the ability to load a high quality version of the photo on the view page. The icon will pulse while the image is loading.

TODO:

* Fix animation codec
* Save animations in proper place with proper names
* Add ability to view animations, and videos for that matter
* Add proper condensing of burst shots
* Add place and landmark recognition
* Speed up scanning somehow?
* Add proper editing tools
* Add search functionality